Refactor to controllers
