+++ controllers/auth/emailConfirmationTemplate.js
