+++ controllers/auth/index.js
@@ -1,0 +1,5 @@
+module.exports = {
+  loginController: require("./loginController"),
+  signupController: require("./signupController"),
+  confirmUserController: require("./confirmUserController")
+};
