+++ controllers/index.js
@@ -1,0 +1,4 @@
+module.exports = {
+  albumController: require("./albumController"),
+  homeController: require("./homeController")
+};
