+++ controllers/upload/index.js
@@ -1,0 +1,4 @@
+module.exports = {
+  getImageController: require("./getImageController"),
+  uploadController: require("./uploadController")
+};
